Fix Exceptions making search timing out
Guzzle ConnectionExceptions would make it so the whole ExpanderPlugin chain would fail because not everything was returned as needed. I encountered this when I was in the Office where the SearchHub endpoints are blocked by our Coorporate Firewall. This braod exception handling makes sure that at least not the whole search breaks in case of errors. "Failing gracefully" basically